# [chore] change JSON marshaler to goccy-json