--- a/cmd/otel-allocator/server/bench_test.go
+++ b/cmd/otel-allocator/server/bench_test.go
@@ -97,8 +97,7 @@ func BenchmarkScrapeConfigsHandler(b *testing.B) {
 func BenchmarkCollectorMapJSONHandler(b *testing.B) {
 	random := rand.New(rand.NewSource(time.Now().UnixNano())) // nolint: gosec
 	s := &Server{
-		logger:         logger,
-		jsonMarshaller: jsonConfig,
+		logger: logger,
 	}
 
 	tests := []struct {
@@ -153,8 +152,7 @@ func BenchmarkCollectorMapJSONHandler(b *testing.B) {
 func BenchmarkTargetItemsJSONHandler(b *testing.B) {
 	random := rand.New(rand.NewSource(time.Now().UnixNano())) // nolint: gosec
 	s := &Server{
-		logger:         logger,
-		jsonMarshaller: jsonConfig,
+		logger: logger,
 	}
 
 	tests := []struct {