**Description:**
jsoniter is no longer maintained; this PR switches over to github.com/goccy/go-json