@atoulme 2025-01-30
Before:
```
GOROOT=/Users/atoulme/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.22.7.darwin-arm64 #gosetup
GOPATH=/Users/atoulme/go #gosetup
/Users/atoulme/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.22.7.darwin-arm64/bin/go test -c -o /Users/atoulme/Library/Caches/JetBrains/IntelliJIdea2024.2/tmp/GoLand/___gobench_github_com_open_telemetry_opentelemetry_operator_cmd_otel_allocator_server.test github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/server #gosetup
/Users/atoulme/Library/Caches/JetBrains/IntelliJIdea2024.2/tmp/GoLand/___gobench_github_com_open_telemetry_opentelemetry_operator_cmd_otel_allocator_server.test -test.v -test.paniconexit0 -test.bench . -test.run ^$
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/server
BenchmarkServerTargetsHandler
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_100
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_100-10         	  359851	      3010 ns/op	    6988 B/op	      25 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_1000
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_1000-10        	  342391	      3052 ns/op	    7008 B/op	      25 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_10000
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_10000-10       	  357673	      2989 ns/op	    7009 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_100000
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_100000-10      	  310776	      3219 ns/op	    7006 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_100
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_100-10        	  357727	      3018 ns/op	    7001 B/op	      25 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_1000
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_1000-10       	  379381	      2964 ns/op	    7009 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_10000
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_10000-10      	  393924	      2960 ns/op	    7009 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_100000
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_100000-10     	  390514	      3266 ns/op	    7007 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_100
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_100-10     	  412264	      2810 ns/op	    6988 B/op	      25 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_1000
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_1000-10    	  391568	      2874 ns/op	    7007 B/op	      25 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_10000
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_10000-10   	  406995	      2920 ns/op	    7009 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_100000
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_100000-10  	  405908	      2964 ns/op	    7006 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_100
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_100-10    	  424369	      2832 ns/op	    7001 B/op	      25 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_1000
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_1000-10   	  374486	      2948 ns/op	    7009 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_10000
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_10000-10  	  366759	      2936 ns/op	    7009 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_100000
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_100000-10 	  318016	      3369 ns/op	    7007 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_100
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_100-10               	  433080	      2848 ns/op	    6980 B/op	      25 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_1000
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_1000-10              	  414410	      2768 ns/op	    6999 B/op	      25 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_10000
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_10000-10             	  422122	      2762 ns/op	    7002 B/op	      25 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_100000
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_100000-10            	  404152	      3053 ns/op	    7000 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_100
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_100-10              	  434769	      3093 ns/op	    6999 B/op	      25 allocs/op
[controller-runtime] log.SetLogger(...) was never called; logs will not be displayed.
Detected at:
	>  goroutine 77 [running]:
	>  runtime/debug.Stack()
	>  	/Users/atoulme/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.22.7.darwin-arm64/src/runtime/debug/stack.go:24 +0x64
	>  sigs.k8s.io/controller-runtime/pkg/log.eventuallyFulfillRoot()
	>  	/Users/atoulme/go/pkg/mod/sigs.k8s.io/controller-runtime@v0.19.3/pkg/log/log.go:60 +0xf4
	>  sigs.k8s.io/controller-runtime/pkg/log.(*delegatingLogSink).WithValues(0x14000aae7c0, {0x140003206a0, 0x2, 0x2})
	>  	/Users/atoulme/go/pkg/mod/sigs.k8s.io/controller-runtime@v0.19.3/pkg/log/deleg.go:168 +0x38
	>  github.com/go-logr/logr.Logger.WithValues(...)
	>  	/Users/atoulme/go/pkg/mod/github.com/go-logr/logr@v1.4.2/logr.go:332
	>  github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/allocation.New({0x1031e36fb, 0x8}, {{0x10493a448?, 0x14000aae7c0?}, 0x3?}, {0x0, 0x0, 0x0})
	>  	/Users/atoulme/w/opentelemetry-operator/cmd/otel-allocator/allocation/strategy.go:89 +0x114
	>  github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/server.BenchmarkServerTargetsHandler(0x14000702a08)
	>  	/Users/atoulme/w/opentelemetry-operator/cmd/otel-allocator/server/bench_test.go:52 +0x354
	>  testing.(*B).runN(0x14000702a08, 0x1)
	>  	/Users/atoulme/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.22.7.darwin-arm64/src/testing/benchmark.go:193 +0x130
	>  testing.(*B).run1.func1()
	>  	/Users/atoulme/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.22.7.darwin-arm64/src/testing/benchmark.go:215 +0x50
	>  created by testing.(*B).run1 in goroutine 1
	>  	/Users/atoulme/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.22.7.darwin-arm64/src/testing/benchmark.go:208 +0x90
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_1000
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_1000-10             	  426286	      2787 ns/op	    7007 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_10000
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_10000-10            	  407527	      2993 ns/op	    7008 B/op	      26 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_100000
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_100000-10           	  385030	      2946 ns/op	    7006 B/op	      26 allocs/op
BenchmarkScrapeConfigsHandler
BenchmarkScrapeConfigsHandler/0_targets
BenchmarkScrapeConfigsHandler/0_targets-10                                        	  512766	      2323 ns/op	    6849 B/op	      26 allocs/op
BenchmarkScrapeConfigsHandler/5_targets
BenchmarkScrapeConfigsHandler/5_targets-10                                        	  434955	      2452 ns/op	    9089 B/op	      26 allocs/op
BenchmarkScrapeConfigsHandler/10_targets
BenchmarkScrapeConfigsHandler/10_targets-10                                       	  391899	      2901 ns/op	   11649 B/op	      26 allocs/op
BenchmarkScrapeConfigsHandler/50_targets
BenchmarkScrapeConfigsHandler/50_targets-10                                       	  275178	      4373 ns/op	   28547 B/op	      26 allocs/op
BenchmarkScrapeConfigsHandler/100_targets
BenchmarkScrapeConfigsHandler/100_targets-10                                      	  165014	      7177 ns/op	   55942 B/op	      26 allocs/op
BenchmarkScrapeConfigsHandler/500_targets
BenchmarkScrapeConfigsHandler/500_targets-10                                      	   54470	     21406 ns/op	  219796 B/op	      26 allocs/op
BenchmarkCollectorMapJSONHandler
BenchmarkCollectorMapJSONHandler/0_collectors_0_targets
BenchmarkCollectorMapJSONHandler/0_collectors_0_targets-10                        	 1455573	       827.9 ns/op	    1752 B/op	      15 allocs/op
BenchmarkCollectorMapJSONHandler/5_collectors_5_targets
BenchmarkCollectorMapJSONHandler/5_collectors_5_targets-10                        	    3062	    386218 ns/op	  535492 B/op	    2792 allocs/op
BenchmarkCollectorMapJSONHandler/5_collectors_50_targets
BenchmarkCollectorMapJSONHandler/5_collectors_50_targets-10                       	     325	   3660315 ns/op	 4991828 B/op	   27622 allocs/op
BenchmarkCollectorMapJSONHandler/5_collectors_500_targets
BenchmarkCollectorMapJSONHandler/5_collectors_500_targets-10                      	      32	  34914491 ns/op	46780893 B/op	  275853 allocs/op
BenchmarkCollectorMapJSONHandler/50_collectors_5_targets
BenchmarkCollectorMapJSONHandler/50_collectors_5_targets-10                       	     318	   3732919 ns/op	 5025731 B/op	   27760 allocs/op
BenchmarkCollectorMapJSONHandler/50_collectors_50_targets
BenchmarkCollectorMapJSONHandler/50_collectors_50_targets-10                      	      32	  34902423 ns/op	46818529 B/op	  276001 allocs/op
BenchmarkCollectorMapJSONHandler/50_collectors_500_targets
BenchmarkCollectorMapJSONHandler/50_collectors_500_targets-10                     	       3	 337770208 ns/op	543216306 B/op	 2757984 allocs/op
BenchmarkCollectorMapJSONHandler/50_collectors_5000_targets
BenchmarkCollectorMapJSONHandler/50_collectors_5000_targets-10                    	       1	3697991084 ns/op	5029540176 B/op	27579637 allocs/op
BenchmarkTargetItemsJSONHandler
BenchmarkTargetItemsJSONHandler/0_targets_0_labels
BenchmarkTargetItemsJSONHandler/0_targets_0_labels-10                             	 1396296	       792.4 ns/op	    1648 B/op	      13 allocs/op
BenchmarkTargetItemsJSONHandler/5_targets_5_labels
BenchmarkTargetItemsJSONHandler/5_targets_5_labels-10                             	  121575	      9617 ns/op	   16868 B/op	      96 allocs/op
BenchmarkTargetItemsJSONHandler/5_targets_50_labels
BenchmarkTargetItemsJSONHandler/5_targets_50_labels-10                            	   15054	     79922 ns/op	  110792 B/op	     567 allocs/op
BenchmarkTargetItemsJSONHandler/50_targets_5_labels
BenchmarkTargetItemsJSONHandler/50_targets_5_labels-10                            	   14005	    107019 ns/op	  153930 B/op	     819 allocs/op
BenchmarkTargetItemsJSONHandler/50_targets_50_labels
BenchmarkTargetItemsJSONHandler/50_targets_50_labels-10                           	    1456	    769680 ns/op	 1071360 B/op	    5537 allocs/op
BenchmarkTargetItemsJSONHandler/500_targets_50_labels
BenchmarkTargetItemsJSONHandler/500_targets_50_labels-10                          	     158	   7427050 ns/op	 9983358 B/op	   55193 allocs/op
BenchmarkTargetItemsJSONHandler/500_targets_500_labels
BenchmarkTargetItemsJSONHandler/500_targets_500_labels-10                         	      14	  93533393 ns/op	109908411 B/op	  514904 allocs/op
BenchmarkTargetItemsJSONHandler/5000_targets_50_labels
BenchmarkTargetItemsJSONHandler/5000_targets_50_labels-10                         	      16	  67886164 ns/op	93553866 B/op	  551638 allocs/op
BenchmarkTargetItemsJSONHandler/5000_targets_500_labels
BenchmarkTargetItemsJSONHandler/5000_targets_500_labels-10                        	       2	 795892542 ns/op	1248932304 B/op	 5148502 allocs/op
PASS

Process finished with the exit code 0
```

After:
```
GOROOT=/Users/atoulme/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.22.7.darwin-arm64 #gosetup
GOPATH=/Users/atoulme/go #gosetup
/Users/atoulme/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.22.7.darwin-arm64/bin/go test -c -o /Users/atoulme/Library/Caches/JetBrains/IntelliJIdea2024.2/tmp/GoLand/___gobench_github_com_open_telemetry_opentelemetry_operator_cmd_otel_allocator_server.test github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/server #gosetup
/Users/atoulme/Library/Caches/JetBrains/IntelliJIdea2024.2/tmp/GoLand/___gobench_github_com_open_telemetry_opentelemetry_operator_cmd_otel_allocator_server.test -test.v -test.paniconexit0 -test.bench . -test.run ^$
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/server
BenchmarkServerTargetsHandler
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_100
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_100-10         	  455734	      2554 ns/op	    6370 B/op	      22 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_1000
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_1000-10        	  414816	      3224 ns/op	    6389 B/op	      22 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_10000
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_10000-10       	  427243	      2654 ns/op	    6393 B/op	      23 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_100000
BenchmarkServerTargetsHandler/least-weighted_num_cols_100_num_jobs_100000-10      	  406468	      4087 ns/op	    6391 B/op	      23 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_100
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_100-10        	  403759	      4557 ns/op	    6383 B/op	      22 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_1000
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_1000-10       	  336313	      3711 ns/op	    6392 B/op	      23 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_10000
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_10000-10      	  351069	      3144 ns/op	    6393 B/op	      23 allocs/op
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_100000
BenchmarkServerTargetsHandler/least-weighted_num_cols_1000_num_jobs_100000-10     	  318456	      3418 ns/op	    6391 B/op	      23 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_100
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_100-10     	  440250	      2761 ns/op	    6370 B/op	      22 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_1000
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_1000-10    	  413475	      2742 ns/op	    6389 B/op	      22 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_10000
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_10000-10   	  377085	      3107 ns/op	    6393 B/op	      23 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_100000
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_100_num_jobs_100000-10  	  353540	      2931 ns/op	    6391 B/op	      23 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_100
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_100-10    	  425239	      2648 ns/op	    6383 B/op	      22 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_1000
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_1000-10   	  424780	      2631 ns/op	    6392 B/op	      23 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_10000
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_10000-10  	  448666	      2697 ns/op	    6393 B/op	      23 allocs/op
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_100000
BenchmarkServerTargetsHandler/consistent-hashing_num_cols_1000_num_jobs_100000-10 	  405319	      2819 ns/op	    6391 B/op	      23 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_100
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_100-10               	  492544	      2401 ns/op	    6363 B/op	      22 allocs/op
[controller-runtime] log.SetLogger(...) was never called; logs will not be displayed.
Detected at:
	>  goroutine 70 [running]:
	>  runtime/debug.Stack()
	>  	/Users/atoulme/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.22.7.darwin-arm64/src/runtime/debug/stack.go:24 +0x64
	>  sigs.k8s.io/controller-runtime/pkg/log.eventuallyFulfillRoot()
	>  	/Users/atoulme/go/pkg/mod/sigs.k8s.io/controller-runtime@v0.19.3/pkg/log/log.go:60 +0xf4
	>  sigs.k8s.io/controller-runtime/pkg/log.(*delegatingLogSink).WithValues(0x1400133d7c0, {0x14006262020, 0x2, 0x2})
	>  	/Users/atoulme/go/pkg/mod/sigs.k8s.io/controller-runtime@v0.19.3/pkg/log/deleg.go:168 +0x38
	>  github.com/go-logr/logr.Logger.WithValues(...)
	>  	/Users/atoulme/go/pkg/mod/github.com/go-logr/logr@v1.4.2/logr.go:332
	>  github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/allocation.New({0x104ee8841, 0x8}, {{0x106668148?, 0x1400133d7c0?}, 0x3?}, {0x0, 0x0, 0x0})
	>  	/Users/atoulme/w/opentelemetry-operator/cmd/otel-allocator/allocation/strategy.go:89 +0x114
	>  github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/server.BenchmarkServerTargetsHandler(0x140015fb688)
	>  	/Users/atoulme/w/opentelemetry-operator/cmd/otel-allocator/server/bench_test.go:52 +0x354
	>  testing.(*B).runN(0x140015fb688, 0x1)
	>  	/Users/atoulme/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.22.7.darwin-arm64/src/testing/benchmark.go:193 +0x130
	>  testing.(*B).run1.func1()
	>  	/Users/atoulme/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.22.7.darwin-arm64/src/testing/benchmark.go:215 +0x50
	>  created by testing.(*B).run1 in goroutine 1
	>  	/Users/atoulme/go/pkg/mod/golang.org/toolchain@v0.0.1-go1.22.7.darwin-arm64/src/testing/benchmark.go:208 +0x90
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_1000
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_1000-10              	  448234	      2481 ns/op	    6382 B/op	      22 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_10000
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_10000-10             	  491791	      2494 ns/op	    6386 B/op	      22 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_100000
BenchmarkServerTargetsHandler/per-node_num_cols_100_num_jobs_100000-10            	  409038	      2635 ns/op	    6385 B/op	      23 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_100
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_100-10              	  462172	      2572 ns/op	    6382 B/op	      22 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_1000
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_1000-10             	  432405	      2524 ns/op	    6391 B/op	      23 allocs/op
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_10000
BenchmarkServerTargetsHandler/per-node_num_cols_1000_num_jobs_10000-10            	  483624	      2542 ns/op	    6393 B/op	      23 allocs/op
…